refactor: move utils to load
